Plant-disease-recognition-in-real-time
In this project, I developed a plant disease recognition system using Keras, a deep learning framework, that can accurately identify healthy plants, plants with powdery mildew, and plants with rust disease. The system utilizes a dataset of images of plants that were collected and labeled by experts in plant pathology.